feat: dynamic permissoins

+++ routes/api.php
@@ -30,31 +30,31 @@ Route::prefix("v1")->group(function () {
     });
 
     Route::group(["prefix" => "roles", "middleware" => ["auth:api"]], function () {
-        Route::get("/", [RoleController::class, "index"]);
-        Route::post("/", [RoleController::class, "store"]);
-        Route::get("/{id}", [RoleController::class, "show"]);
-        Route::put("/{id}", [RoleController::class, "update"]);
-        Route::delete("/{id}", [RoleController::class, "destroy"]);
+        Route::get("/", [RoleController::class, "index"])->middleware('perm:roles.view');
+        Route::post("/", [RoleController::class, "store"])->middleware('perm:roles.create');
+        Route::get("/{id}", [RoleController::class, "show"])->middleware('perm:roles.view');
+        Route::put("/{id}", [RoleController::class, "update"])->middleware('perm:roles.update');
+        Route::delete("/{id}", [RoleController::class, "destroy"])->middleware('perm:roles.delete');
     });
 
     Route::group(["prefix" => "permissions", "middleware" => ["auth:api"]], function () {
-        Route::get("/", [PermissionController::class, "index"]);
+        Route::get("/", [PermissionController::class, "index"])->middleware('perm:permissions.view');
     });
 
     Route::group(["prefix" => "users", "middleware" => ["auth:api"]], function () {
-        Route::get("/", [UserController::class, "index"]);
-        Route::post("/", [UserController::class, "store"]);
-        Route::get("/{id}", [UserController::class, "show"]);
-        Route::put("/{id}", [UserController::class, "update"]);
-        Route::delete("/{id}", [UserController::class, "destroy"]);
+        Route::get("/", [UserController::class, "index"])->middleware('perm:users.view');
+        Route::post("/", [UserController::class, "store"])->middleware('perm:users.create');
+        Route::get("/{id}", [UserController::class, "show"])->middleware('perm:users.view');
+        Route::put("/{id}", [UserController::class, "update"])->middleware('perm:users.update');
+        Route::delete("/{id}", [UserController::class, "destroy"])->middleware('perm:users.delete');
     });
 
-    Route::middleware(['auth:api'])->group(function () {
-        Route::get('holidays', [HolidayController::class, 'index']);
-        Route::get('holidays/{id}', [HolidayController::class, 'show']);
-        Route::post('holidays', [HolidayController::class, 'store']);
-        Route::put('holidays/{id}', [HolidayController::class, 'update']);
-        Route::delete('holidays/{id}', [HolidayController::class, 'destroy']);
-        Route::post('holidays/bulk-import', [HolidayController::class, 'bulkImport']);
+    Route::middleware(["prefix" => "holidays", 'auth:api'])->group(function () {
+        Route::get('/', [HolidayController::class, 'index'])->middleware('perm:holidays.view');
+        Route::post('/', [HolidayController::class, 'store'])->middleware('perm:holidays.create');
+        Route::get('/{id}', [HolidayController::class, 'show'])->middleware('perm:holidays.view');
+        Route::put('/{id}', [HolidayController::class, 'update'])->middleware('perm:holidays.update');
+        Route::delete('/{id}', [HolidayController::class, 'destroy'])->middleware('perm:holidays.delete');
+        Route::post('/bulk-import', [HolidayController::class, 'bulkImport'])->middleware('perm:holidays.create');
     });
 });
